+++ web/src/contexts/project-page-context.tsx
@@ -2,11 +2,6 @@ import React, { createContext, useContext, useEffect, useState } from 'react';
 import { useParams, useSearchParams } from 'react-router-dom';
 
 import { useProjectAnnotation } from '../hooks/queries/useProjectAnnotation';
-import { useProjectConfig } from '../hooks/queries/useProjectConfig';
-import { useProjectViews } from '../hooks/queries/useProjectViews';
-import { useSampleTable } from '../hooks/queries/useSampleTable';
-import { useSubsampleTable } from '../hooks/queries/useSubsampleTable';
-import { useValidation } from '../hooks/queries/useValidation';
 
 const MAX_SAMPLE_COUNT = 25_000;
 
@@ -18,11 +13,6 @@ const ProjectPageContext = createContext<{
   namespace: string;
   projectName: string;
   tag: string;
-  sampleTableQuery?: ReturnType<typeof useSampleTable>;
-  subSampleTableQuery: ReturnType<typeof useSubsampleTable>;
-  projectConfigQuery: ReturnType<typeof useProjectConfig>;
-  projectViewsQuery: ReturnType<typeof useProjectViews>;
-  projectValidationQuery: ReturnType<typeof useValidation>;
   shouldFetchSampleTable: boolean;
   forceTraditionalInterface: boolean;
   setForceTraditionalInterface: React.Dispatch<React.SetStateAction<boolean>>;
@@ -56,43 +46,12 @@ export const ProjectPageProvider = ({ children }: ProviderProps) => {
     shouldFetchSampleTable = projectAnnotationQuery.data.number_of_samples <= MAX_SAMPLE_COUNT;
   }
 
-  // SAMPLE TABLE
-  const sampleTableQuery = useSampleTable({
-    namespace,
-    project: projectName,
-    tag,
-    enabled: projectAnnotationQuery.data === undefined ? false : shouldFetchSampleTable,
-  });
-
-  // SUBSAMPLE TABLE
-  const subSampleTableQuery = useSubsampleTable(namespace, projectName, tag);
-
-  // PROJECT CONFIG
-  const projectConfigQuery = useProjectConfig(namespace, projectName, tag);
-
-  // PROJECT VIEWS
-  const projectViewsQuery = useProjectViews(namespace, projectName, tag);
-
-  // PROJECT VALIDATION
-  const projectValidationQuery = useValidation({
-    pepRegistry: `${namespace}/${projectName}:${tag}`,
-    schema: projectAnnotationQuery.data?.pep_schema || 'pep/2.0.0', // default to basic pep 2.0.0 schema
-    schema_registry: projectAnnotationQuery.data?.pep_schema,
-    enabled:
-      namespace && projectName && tag && projectAnnotationQuery.data === undefined ? false : shouldFetchSampleTable,
-  });
-
   return (
     <ProjectPageContext.Provider
       value={{
         namespace,
         projectName,
         tag,
-        sampleTableQuery,
-        subSampleTableQuery,
-        projectConfigQuery,
-        projectViewsQuery,
-        projectValidationQuery,
         shouldFetchSampleTable,
         forceTraditionalInterface,
         setForceTraditionalInterface,
